Add a property 'username' to User
This property becomes the primary key instead of the email.
The reason behind this change is that we need a simple unique identifier for each user that can be used in the routes of the API. We could have used a classic numeric identifier, but the calls to the API would be less self-talking.